requirements.txt
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
idna==3.10
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
structlog==25.4.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
typer[all]
openai
pydantic-settings
loguru
psutil
jinja2
